queue traits to its own file
