+++ libs/mocks/src/lib.rs
@@ -10,11 +10,11 @@ pub mod foreign_investment_hooks;
 pub mod investment;
 pub mod liquidity_pools;
 pub mod liquidity_pools_gateway;
-pub mod liquidity_pools_gateway_queue;
 pub mod pay_fee;
 pub mod permissions;
 pub mod pools;
 pub mod pre_conditions;
+pub mod queue;
 pub mod rewards;
 pub mod router_message;
 pub mod status_notification;
@@ -30,7 +30,6 @@ pub use fees::pallet as pallet_mock_fees;
 pub use investment::pallet as pallet_mock_investment;
 pub use liquidity_pools::pallet as pallet_mock_liquidity_pools;
 pub use liquidity_pools_gateway::pallet as pallet_mock_liquidity_pools_gateway;
-pub use liquidity_pools_gateway_queue::pallet as pallet_mock_liquidity_pools_gateway_queue;
 pub use pay_fee::pallet as pallet_mock_pay_fee;
 pub use permissions::pallet as pallet_mock_permissions;
 pub use pools::pallet as pallet_mock_pools;
