+++ libs/traits/src/lib.rs
@@ -30,25 +30,16 @@ use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
 use sp_runtime::{traits::Member, DispatchError};
 use sp_std::{fmt::Debug, marker::PhantomData, vec::Vec};
 
-/// Traits related to checked changes.
 pub mod changes;
-/// Traits related to data registry and collection.
 pub mod data;
-/// Traits related to Ethereum/EVM.
 pub mod ethereum;
-/// Traits related to pool fees.
 pub mod fee;
-/// Traits related to fees payment.
 pub mod fees;
-/// Traits related to interest rates.
 pub mod interest;
-/// Traits related to investments.
 pub mod investments;
-/// Traits related to liquidity pools.
 pub mod liquidity_pools;
-/// Traits related to rewards.
+pub mod queue;
 pub mod rewards;
-/// Traits related to swaps.
 pub mod swaps;
 
 #[cfg(feature = "runtime-benchmarks")]
