+++ libs/traits/src/liquidity_pools.rs
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use frame_support::{dispatch::DispatchResult, weights::Weight};
+use frame_support::dispatch::DispatchResult;
 use sp_runtime::DispatchError;
 use sp_std::vec::Vec;
 
@@ -76,27 +76,6 @@ pub trait MessageReceiver {
 		origin: Self::Origin,
 		message: Vec<u8>,
 	) -> DispatchResult;
-}
-
-/// The trait required for queueing messages.
-pub trait MessageQueue {
-	/// The message type.
-	type Message;
-
-	/// Submit a message to the queue.
-	fn submit(msg: Self::Message) -> DispatchResult;
-}
-
-/// The trait required for processing queued messages.
-pub trait MessageProcessor {
-	/// The message type.
-	type Message;
-
-	/// Process a message.
-	fn process(msg: Self::Message) -> (DispatchResult, Weight);
-
-	/// Process a message.
-	fn max_processing_weight(msg: &Self::Message) -> Weight;
 }
 
 /// The trait required for handling outbound LP messages.
