+++ pallets/liquidity-pools-gateway/src/message_processing.rs
@@ -1,5 +1,6 @@
-use cfg_traits::liquidity_pools::{
-	InboundMessageHandler, LPEncoding, MessageQueue, Proof, RouterProvider,
+use cfg_traits::{
+	liquidity_pools::{InboundMessageHandler, LPEncoding, Proof, RouterProvider},
+	queue::MessageQueue,
 };
 use cfg_types::domain_address::{Domain, DomainAddress};
 use frame_support::{
