+++ pallets/liquidity-pools-gateway/src/tests.rs
@@ -1,7 +1,10 @@
 use std::collections::HashMap;
 
 use cfg_primitives::LP_DEFENSIVE_WEIGHT;
-use cfg_traits::liquidity_pools::{LPEncoding, MessageProcessor, OutboundMessageHandler};
+use cfg_traits::{
+	liquidity_pools::{LPEncoding, OutboundMessageHandler},
+	queue::MessageProcessor,
+};
 use cfg_types::domain_address::*;
 use frame_support::{assert_err, assert_noop, assert_ok};
 use itertools::Itertools;
