+++ runtime/integration-tests/src/cases.rs
@@ -6,11 +6,11 @@ mod ethereum_transaction;
 mod example;
 mod investments;
 mod liquidity_pools;
-mod liquidity_pools_gateway_queue;
 mod loans;
 mod lp;
 mod oracles;
 mod proxy;
+mod queue;
 mod restricted_transfers;
 mod routers;
 mod xcm_transfers;
